+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## master
 * Add TODO
 * Add abstract class Aihrus_Settings
+* Add is_true, terms, url validate cases
+* Add shortcode id helpers
 * Settings link with null post_type
 
 ## 0.0.0